M2M (`Todo` and `Category`) was switched to M2O: improved only logic, no ui yet

+++ data/src/main/java/com/example/data/entity/TodoEntity.kt
@@ -1,9 +1,10 @@
 package com.example.data.entity
 
 import androidx.room.ColumnInfo
+import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.example.domain.model.Todo
+import com.example.domain.model.Category
 import java.util.*
 
 @Entity(tableName = "todos")
@@ -12,7 +13,7 @@ data class TodoEntity(
     val id: Long = 0,
     val title: String,
     @ColumnInfo(name = "is_completed") val isCompleted: Boolean,
-    @ColumnInfo(name = "category_id") val categoryId: Int?,
+    @ColumnInfo(name = "category_id") val categoryId: Long?,
     @ColumnInfo(name = "deadline_date") val deadlineDate: Date?
 ) {
 }
